README.md: markdown-ify code blocks, links & lists
For markdown to treat indented blocks of text as code snippets it is
necessary to separate the code snippets from regular text with a
newline.

Likewise lists must use the format "^ * $text" to create bullet points
for list items.

This patch properly indents text blocks and bullet points to existing
text and converts plainly printed URLs into hyperlinks with alt-text.

Signed-off-by: Mike Turquette <mturquette@linaro.org>