fixed segmentation fault when calling "audit perf" command
In this patch added an initialization pointers "time_in_opp_t0",
"time_in_opp_t1", "time_in_opp_cnt" with value NULL at the beginning
of the function "audit_performances".
Because if the function "cpufreq_opp_nbr_get" returns 0
(file ".../time_in_state" is absent), then the memory for them
is not allocated (pointers are remain uninitialized) and attempt
to release allocated memory for them in the end of the function
"audit_performances" results in segmentation fault.

Signed-off-by: Oleksandr_Tyshchenko <oleksandr.tyshchenko@ti.com>