fix empty interrupt statistic when CPU1 is offline
File "/proc/interrupts" contains statistics for CPUs that are not
only available, but also online. That is, when reading the
information from this file should use the number of CPUs that are
online.
In this patch, in all functions where the information is read from
the file "/proc/interrupts", function "cpu_cores_count_get"
is replaced by the function "cpu_online_cores_count_get".

Signed-off-by: Oleksandr_Tyshchenko <oleksandr.tyshchenko@ti.com>