[OMAP4][OPP] fix OPP-clock rates incoherency and highlight disabled modules
Fix potential OPP-clock rates incoherency that could be observed using "show opp" command.
This was due to the fact that OPP could change in backgroup, while voltdm44xx_opp_show() was
sampling sequencially all clock rates.
Method: sample all parameters in a loop until all is consistent.

If a module is disabled, printing its clock rate is not relevant, and may be misleading.
Add parenthesis about rates for disabled modules as well as a note at the footer of the table.

Signed-off-by: Patrick Titiano <p-titiano@ti.com>