[OMAP4460][GALAXY NEXUS][AUDIT][SYSCONFIG][HACK] skip DSS modules when DSS domain is not ON
On Galaxy Nexus phone with release Jelly Bean JOP40D (not tested on other release), when display
is OFF and DSS power domain is RET, DISPC module is actually left in module mode 0x2 ("explicitly enabled").
This is tolerated by HW (...) but confuses omapconf, as it believes DSS sub-modules like DSI[1-2]/HDMI/RFBI/...
are accessible.
This causes omapconf to attempt to read a register that is not clocked and it crashes.
Adding this hack to workaround this problem.

Signed-off-by: Patrick Titiano <p-titiano@ti.com>