cpuinfo: OMAP5: fix cpu device type detection
OMAP5 ES2.0 has a different register compared to OMAP4 for
CONTROL_STATUS.

In addition, the bit range used to store the type of the device has also
been changed w.r.t OMAP4.

Current code reads from the wrong CONTROL register and wrong bits, as a
result of which CPU type is wrongly displayed. Update the code for the
same.

Signed-off-by: Jin Zheng <j-zheng@ti.com>