Makefile: allow to build tags and cscope
It is handy to have cscope and tags built from the sources to
account for any change in source without having to run special
scripts
So, add option to 'make tags' and 'make cscope' to generate
the references.

While at it, ignore these generated files from git tracking.

Note: we use -MM -MG options of gcc for this to work to ensure that
we can use the header dependency even for files that have not been
committed to git repository

[j-zheng@ti.com: update to keep in sync with latest code]
Signed-off-by: Jin Zheng <j-zheng@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>