[PMIC] use array index matching for PMIC matches
In original code, the pmic_names and smps_names arrays
are initialized base upon the enum values in pmic_id and
pmic_smps_id.

As we keep adding new PMIC and voltage domains, it'll increase
the mismatching chances. Therefore, it'll be handy if we initialize
the arrays by matching the values exactly to its corresponding index.

Acked-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Jin Zheng <j-zheng@ti.com>