[DRA7][PRCM] Use SYS_CLK1 rates based on CM_CLKSEL_SYS
The SYS_CLK1 rate is currently hard-coded to 20 MHz matching
the value on the DRA7 EVM, but not every one will be using
the same crystal frequency. Use the CM_CLKSEL_SYS register
value to compute the SYS_CLK1 rate. The rates are limited
to 19.2, 20 and 27 MHz as per the latest DM manual.

NOTE: SYS_CLK2 rate is also hard-coded similarly, but there
doesn't seem to be a register for reading it. So, for accuracy,
this value needs to be updated in code.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Nishant Menon <nm@ti.com>