[CPUINFO][AM335X] Add support for AM335X SoC
AM335X is now supported on cpuinfo. This required the ability to
differentiate between AM335X and OMAP chips, which was solved by
reading the machine file of soc0. Package type was determined through
the EFUSE_SMA register (silicon revisions 2.0 and 2.1). For silicon
revision 1.0, package type is left as unknown. Silicon revision was
determined by DEVICE_ID register, and chip type was determined by
DEV_FEATURE register.

There is a small comment on the AM335X DIEID in the cpu get DIEID
function. The AM335X DIEID is TI proprietary information and cannot be
displayed.

While at it, add the necessary components for the project to build
with the AM335X files and file paths.

Signed-off-by: Gilberto Rodriguez <gilrod@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>