[AM335X][DPLL] Add definitions and helper to enable show dpll
Add more definitions and structs needed by the omapconf show dpll
command. Define basic clocks for the DPLL and a few system clocks.
Include functions to calculate clock rate for the Master Oscillator
(SYS_CLK) and define the clock rates for the different DPLL and system
clocks.

Add functions to support the show dpll command. DPLL init and free are
implemented to gather the different DPLL settings and remove them when
the settings are no longer needed. Add the settings extract function
for both the DPLL and hsdivs. Calculate the different output rates as
well as gather the necessary clock output rates.

Signed-off-by: Gilberto Rodriguez <gilrod@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>