[AM335X][MODULES] Add AM335X modules for show opp
Add a few AM335X modules to support the show opp command, i.e., l3 and
l4 interconnects, emif, and mpu modules. Defined most of the mods that
could be implemented in the future. Update the common module files to
support the AM335X module functions.

Include the neccessary defines in the common opp header to build.

Signed-off-by: Gilberto Rodriguez <gilrod@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>