[AM335X][LIB] Add support for Register name based read/write
Add name2addr functions for CM and PRM and add a lib for AM335X with
basic functions to allow 'omapconf read/write' to use the register
names as listed in the TRM.

Caveat: This only supports PRCM registers for now. Update common lib
to call to AM335X find register function.

While at it, fix a minor formatting error for prm header.

Signed-off-by: Gilberto Rodriguez <gilrod@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>