[DRA7/COMMON][STATCOLL] Fix conflict between probe and master option selection for ma_mpu
When -m ma_mpu -p emif1 is selected, the channel retains alldmm
(default) as the master. This is wrong since we explicitly requested
for mpu as the master for the filter configuration.

This occurs because there are only checks for incorrect MA_MPU probe
channel configurations, and given none, it assumes a non-MA_MPU
config. It needs a check to retain correct MA_MPU configurations.

Added the missing checks to fix the issue.

Signed-off-by: Edward Shin <e-shin@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>