[DRA7][CROSSBAR] Add PRU-ICSS interrupt info to crossbar data
AM57xx SoCs have two PRU-ICSS subsystems, each of which has 8
interrupts routed to the Crossbar. Add the interrupt information
to the common DRA7 Crossbar data, so that this shows up when
doing the crossbar dump.

NOTE:
The sub-systems are not supported on the DRA7 SoCs, so added a
"AM57xx only" to the interrupt information, as the corresponding
interrupts are marked reserved in the DRA7 TRMs.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>