[DRA7][CROSSBAR] Add support for PRU-ICSS in crossbar dump and audit
AM57xx SoCs have two PRU-ICSS subsystems, each of which supports
64 input events routed to their respective INTCs. Events 32 to 63
for each PRU-ICSS subsystem are also routed through a Crossbar.

This patch extends the crossbar dump and audit support to the two
PRU-ICSS Crossbar instances.

NOTE:
The sub-systems are not supported on the DRA7 SoCs, so added a
small note to the help text.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>