[CPUINFO] Correct I2C bus number to fix cpuinfo for OMAP4/5
The cpuinfo option prints out the PMIC and AudioIC device
information alongside the SoC information. This data is
gathered by querying the devices using appropriate i2c
commands. The I2C bus number being used currently matches
the bus numbers from the legacy non-DT boots, and does not
match the DT alias numbers from the current DT-based boots.
The PMIC and AudioIC info is incorrect as a result, and it
also causes additional delays due to i2c timeouts on every
omapconf command (especially on OMAP5).

Fix this I2C bus number so that the cpuinfo information is
printed properly for OMAP4 and OMAP5 SoCs. The behavior on
other SoCs is same as before this patch.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>