[DRA7] Add support for IPU AMMU decode
The IPU subsystem of the AM57xx and DRA7xx processors contains an
Attribute MMU (AMMU). It sometimes is called the "Unicache MMU"
or the "Shared Cache MMU". This patch is intended to decode the
address mappings and associated flags. The primary reference for
this patch set was the AM57xx TRM, SPRUHZ6 Rev H, Chapter 7.4.4
"IPUx_UNICACHE_MMU (AMMU) Registers". However, other TRMs, like
the OMAP4470 Public TRM (SWPU270 Rev T) were also consulted in
order to make the code as reusable as possible.

I chose to structure these patches as having a "common" area plus
a device-specific area since there are other devices that also
implement the AMMU. For example, OMAP4 and OMAP5 also use the
Attribute MMU, but they have a different number of instances of
the IPU (Cortex M3/M4 subsystem). Also, OMAP4 and OMAP5 contained
an AMMU for their DSP subsystem since they used a different DSP
core. So if desired it would be fairly simple for someone to create
a follow-up patch that adds OMAP4/5 support where they could leverage
the underlying register decoding that is common across devices.

Signed-off-by: Brad Griffis <bgriffis@ti.com>
[nm@ti.com: minor cleanups]
Signed-off-by: Nishanth Menon <nm@ti.com>