[AM4][CTT] Align register reporting with other tools
Register set is updated to match the AM43xx variant.  This is
a superset of all AM43xx, i.e. it encompasses AM437x, AM437xHS
and AM438x.

This also aligns with other tools such as this DSS file:

https://git.ti.com/sitara-dss-files/am43xx-dss-files/blobs/master/am43xx-ctt.dss

In particular, the JTAG script is precisely aligned with the
formatting of omapconf such that you can dump the registers with
either one and compare the result without picking up little differences
like the case of the hexadecimal, missing registers, etc.

Signed-off-by: Brad Griffis <bgriffis@ti.com>